feat: add IPreSaleOrchestrator_v1 interface and update project configuration

+++ test/ERC20Issuance_v1.t.sol
@@ -2,11 +2,11 @@
 pragma solidity ^0.8.0;
 
 import "forge-std/Test.sol";
-import {ERC20Issuance_v1} from "@/ERC20Issuance_v1.sol";
+import {ERC20Issuance_v1} from "src/ERC20Issuance_v1.sol";
 
 contract ERC20Issuance_v1Test is Test {
     // CONSTANTS
-    uint constant MAX_SUPPLY = type(uint256).max - 1;
+    uint256 constant MAX_SUPPLY = type(uint256).max - 1;
     string constant NAME = "MyToken";
     string constant SYMBOL = "MTK";
     uint8 constant DECIMALS = 18;
@@ -17,30 +17,20 @@ contract ERC20Issuance_v1Test is Test {
 
     function setUp() public {
         // Deploy the MyToken contract
-        token = new ERC20Issuance_v1(
-            NAME,
-            SYMBOL,
-            DECIMALS,
-            MAX_SUPPLY,
-            deployer
-        );
+        token = new ERC20Issuance_v1(NAME, SYMBOL, DECIMALS, MAX_SUPPLY, deployer);
     }
 
     function testMint() public {
         // Set the mint amount
-        uint mintAmount = 1000;
+        uint256 mintAmount = 1000;
 
         // Mint the tokens to the deployer's address
         token.mint(deployer, mintAmount);
 
         // Check that the initial supply was allocated to the deployer's address
-        uint deployerBalance = token.balanceOf(deployer);
+        uint256 deployerBalance = token.balanceOf(deployer);
 
-        assertEq(
-            deployerBalance,
-            mintAmount,
-            "Deployer should have the minted amount"
-        );
+        assertEq(deployerBalance, mintAmount, "Deployer should have the minted amount");
     }
 
     function testTokenDetails() public view {
